Rework download to be more testable
Separate starting the download from initializing the Download object,
and join the download thread before setting expectations on the outcome
of the test.

+++ spec/shoes/download_spec.rb
@@ -21,37 +21,43 @@ describe Shoes::Download do
   end
 
   after do
-    download.join_thread
-    File.delete opts[:save]
+    File.delete opts[:save] if File.exist? opts[:save]
   end
 
-  it "finishes" do
-    eventually { expect(download).to be_finished }
+  context 'before it has been started' do
+    it 'understands percent' do
+      eventually{ expect(percent).to eql(0)}
+    end
+
+    it 'understands abort' do
+      expect(download).to respond_to(:abort)
+    end
+
+    it 'understands content_length' do
+      expect(content_length).to be >= 1
+    end
+
+    it 'understands length' do
+      expect(length).to be >= 1
+    end
   end
 
-  it 'starts' do
-    eventually { expect(download).to be_started }
-  end
+  context 'without a progress proc' do
+    before :each do
+      download.start
+    end
 
-  it 'understands percent' do
-    eventually{ expect(percent).to eql(0)}
-  end
+    it 'finishes' do
+      eventually { expect(download).to be_finished }
+    end
 
-  it 'understands abort' do
-    expect(download).to respond_to(:abort)
-  end
+    it 'starts' do
+      eventually { expect(download).to be_started }
+    end
 
-  it "understands content_length" do
-    expect(content_length).to be >= 1
-  end
-
-  it "understands length" do
-    expect(length).to be >= 1
-  end
-
-  it 'creates the file specified by save' do
-    download
-    eventually { expect(File.exist?(opts[:save])).to be_truthy }
+    it 'creates the file specified by save' do
+      eventually { expect(File.exist?(opts[:save])).to be_truthy }
+    end
   end
 
   context 'with a progress proc' do
@@ -59,29 +65,27 @@ describe Shoes::Download do
     let(:opts) { {save: "nasa50th.gif", progress: progress_proc} }
     subject(:download) { Shoes::Download.new app, parent, name, opts}
 
-    it 'calls the progress proc from start, download and finish' do
-      skip 'Fails randomly.. maybe investigate with #681'
+    before :each do
       allow(download.gui).to receive :eval_block
-      eventually {
+      download.start
+      download.join_thread
+    end
+
+    context 'with content length' do
+      it 'calls the progress proc from start, download and finish' do
         expect(download.gui).to have_received(:eval_block).
                                   with(progress_proc, download).
                                   exactly(3).times
-      }
+      end
     end
 
     context 'without content length' do
-      before do
-        stub_request(:get, name)
-          .to_return(:headers => {}, :status => response_status, :body => response_body)
-      end
+      let(:response_headers) { Hash.new }
 
-      it 'does not fail on progress, but called from content length and finish' do
-        skip 'Sometimes fails on Travis'
-        eventually {
-          expect(download.gui).to receive(:eval_block).
+      it 'does not call on progress, but called from content length and finish' do
+        expect(download.gui).to have_received(:eval_block).
           with(progress_proc, download).
           twice
-        }
       end
     end
   end
@@ -90,25 +94,31 @@ describe Shoes::Download do
     let(:result) {download}
     let(:response) {download.response}
 
+    before :each do
+      allow(download.gui).to receive(:eval_block)
+      download.start
+      download.join_thread
+    end
+
     context 'with a block' do
 
       it 'calls the block with a result' do
-        skip 'damn you download specs we really need to you to be reliable'
+        #skip 'damn you download specs we really need to you to be reliable'
         # https://travis-ci.org/shoes/shoes4/jobs/25269033
-        eventually { expect(download.gui).to receive(:eval_block).with(input_block, result) }
+        expect(download.gui).to have_received(:eval_block).with(input_block, result)
       end
 
       describe 'response object' do
         it 'has headers hash' do
-          eventually { expect(response.headers).to eql(response_headers)}
+          expect(response.headers).to eql(response_headers)
         end
 
         it 'has body text' do
-          eventually { expect(response.body).to eql(response_body)}
+          expect(response.body).to eql(response_body)
         end
 
         it 'has status array' do
-          eventually { expect(response.status).to eql(response_status)}
+          expect(response.status).to eql(response_status)
         end
       end
     end
@@ -117,7 +127,7 @@ describe Shoes::Download do
       subject(:download) { Shoes::Download.new app, parent, name, opts }
 
       it 'does not call the block' do
-        expect(download.gui).not_to receive(:eval_block)
+        expect(download.gui).not_to have_received(:eval_block)
       end
     end
 
@@ -127,8 +137,8 @@ describe Shoes::Download do
       subject(:download) { Shoes::Download.new app, parent, name, opts}
 
       it 'calls the finish proc' do
-        skip 'Another Travis failure...'
-        eventually { expect(download.gui).to receive(:eval_block).with(finish_proc, result) }
+        #skip 'Another Travis failure...'
+        expect(download.gui).to have_received(:eval_block).with(finish_proc, result)
       end
     end
